Cleanup for perf and move stuff into Practice
